chore(deps): setup typescript in backend
